Add special tracking behaviour for the /ask transaction
The gov.uk/ask service cannot use the existing cross-domain tracking
abilities set-up in publisher as we do not have google analytics running
on the destination Smart Survey, instead we need to pass this via a
query string variable.

To achieve this we look for any links to our particular Smart Survey
when the page loads for users who have allowed Google Analytics and then
add a query string of the client Id from Analytics.

This change ports over the work from https://github.com/alphagov/government-frontend/pull/1744
which was done to enable this to run on an Answer page format, moving it
here allows this to be done on a transaction format instead.

I've put in the famous last words in a comment of "this should only be
short term" hopefully this doesn't come back to haunt me.