Use view context for contextual URLs
Previously this used the approach of using Plek as a short cut to
getting an asset_url in each environment. This was somewhat dubious as
it relied on asset_url actually returning a path rather than a full URL
and was dependent on the Plek asset_root matching what was configured in
the app as the asset_host.

To resolve this we pass in the view_context from the controller to the
presenter, this allows generating a URL based off a combination of the
request context and the Rails app configuration.