Make outcome body safe for simple smart answers
This was changed in 5fb281fc560b9a1d77a27ff555c834dcd1c7df09 and it
prevents tables from rendering correctly on the pages:

See
https://www.gov.uk/student-finance-forms/y/english-student-part-time/apply-for-student-loans-and-grants/2018-to-2019/no/no/yes
for an example.

I think it's fine to mark this string as HTML safe because it comes from
a trusted source and can only be modified by a content designer.