Add format constraint for simple smart answer routing
When a simple smart answer document is replaced with a transaction page,
users may still visit paths for the smart answer.

This is currently serving a 500 error and logging an error to Sentry.

Adding a format constraint so we only try to serve simple smart answers
that are actually documents of that format.  A 404 error will be served
where the smart answer has been removed.