Update smartsurvey host name and path
We now use a GOV.UK hostname for our surveys so this is updated
accordingly, we also use the URL without a trailing slash.

At the time of writing there isn't much tracking of user journeys for
smartsurvey usage, however the team decided they wanted to keep this
functionality in place for the time being while surveys were still being
run so that a dashboard could be re-opened.