Update simple smart answers headings
The headings for each step of a simple smart answer were the publication's
title. This meant that the heading on each step of the simple smart answer was
the same, which could cause a screen reader user to think that the page hadn't
progressed to the next step.

The headings for each step of the simple smart answer have been changed to the
question itself, with the name of the simple smart answer itself displayed as
the context of the title. The heading size has been reduced from "xl" to "l".