Remove and reorder smart answer flow pages
The design of the page requires the summary of the answered questions at the
top of the page, with the question below it. If the markup was done this way
then this would lead to either multiple `h1`s (not good) or a `h2` preceding a
`h1` (not allowed).

To prevent this the markup has the question first, then the summary below; and
the visual layout has the summary first and the question below.

This still passes WCAG SC 1.3.2 Meaningful Sequence as the guidance states
that ["providing a particular linear order is only required where it affects
meaning"](https://www.w3.org/TR/UNDERSTANDING-WCAG20/content-structure-separation-sequence.html).
The visual order of the sections doesn't affect the meaningful sequence as
they are as easily understood in whichever order they come in.