Change smart answer page layout
The previous design of the page requires the summary of the answered questions
at the top of the page, with the question below it. If the markup was done
this way then this would lead to either multiple `h1`s (not good) or a `h2`
preceding a `h1` (also not good).

The previous design of the page also didn't match the smart answers layout,
which has the question first and the summary afterwards.

The simple smart answers now has the question first and the summary second in
both the markup and the visual layout. The questions have the question as the
`h1`, with the title of the simple smart answer as the subtitle.

The outcome page also has the outcome as the title, with the name of the simple
smart answer as the subtitle.