Changes page `<title>` logic
The problem - simple smart answers were displaying the same text in each
step's `<title>`, which means that the pages don't meet the [success criteria
2.4.2 of the WCAG](https://www.w3.org/WAI/WCAG21/Understanding/page-titled.html).

The page title has an conditional statment where if the page is a publication
the `<title>` is set by a helper. This means that using `content_for :title`
doesn't work for publications because the title text is generated by the
helper - and simple smart answers are considered a publication.

This pull request flips the logic - `:title` is used if present, otherwise the
publication title helper is used, and finally a fallback of "GOV.UK". This
allows the `<title>` of a page to be set from with the template, even if the
page is a publication - which means we can fix the `<title>` for the simple
smart answers.