Add manifest for PaaS application
This manifest file 'tells Cloud Foundry what to do with your app'.

It is equivalent to a task definition in AWS ECS.

To create the app one must sign into PaaS and run the following
commands:

cf push
cf set-env <app_name> SECRET_KEY_BASE string
cf set-env <app_name> BASIC_AUTH_PASSWORD string

This is sufficient to get an application running at
https://frontendreplatforming.cloudapps.digital/

It is also possible to deploy an app using a Docker
image. However, we would be responsible for ensuring
the security of the Docker image. By using a buildpack
(a standard one) we are not responsible for the build
environment or the operating system.
Details: https://docs.cloud.service.gov.uk/responsibility_model.html#know-your-responsibilities