Fix HTML render issues
Issue: HTML was not being rendered correctly in 'govspeak' blocks.

Solution: It was found that the 'sanitize' method was being used.
This method strips all HTML tags and attributes that aren’t whitelisted.
This means that the content being passed into the 'govspeak' blocks was
plain text.

We need to ensure the HTML presented is safe, so applying the
'html_safe' method.  Unfortunately, some fields are not mandatory
so we need guard clauses to stop 'nil' exceptions on all optional
fields.