Remove redundant GitHub Actions config
Previously we added this as part of a migration away from Jenkins,
but this has stalled because Jenkins is still required for some of
the tasks (e.g. release tagging), and so now we have both. Since
the build is currently locked on a stuck action, I'd prefer to
remove this redundant bit of CI than try to fix it.