Run Jasmine tests with selenium driven Chrome
This changes the configuration mechanism for Jasmine to use headless
Chrome rather than PhantomJS. We need to govuk_test as a development
dependency as the files need to be available when loading files for
the Jasmine rake task.

This also removes the conditional that only added all.js and
test_dependencies.js in the test environment. This caused problems when
running a Jasmine webserver (presumably that compiles assets in a dev
environment). Removing this is of low consequence as these files are
merely made available they're not forced on the app.