Change aria-live region to one that's less verbose
Previously the entire list of countries was being read out as you typed into the
search input. This is too verbose and led to the page being nearly unusable
when using JAWS.

For example, if you were searching for "Australia" all the countries with "A" in
their name would be read out (all 198 of them), then all the countries with
"Au" (9), then the two with "Aus" in them.

This has been changed to only read out the number of countries or territories
that found for a search result. This means that searching for "Australia" will
cause "198 countries or territories" to be read out, then "9 countries or
territories", then "Two countries or territories".