Use GdsApi rather than Services
We don't need to keep our own instance of the GdsApi class as we can use
`GdsApi` directly. This follows on from
https://github.com/alphagov/frontend/pull/2013 which seems to have
returned when we merged calendars into the app.

The tests should probably be updated to use the stub methods provided by
gds-api-adapters but I didn't want to change too much in this PR.