Add CachedContentItem
This caches the response from `GdsApi.content_store.content_item` for 30
seconds. I've updated the various places where
`GdsApi.content_store.content_item` is used to call
`CachedContentItem.fetch` instead.

This should help us to handle spikes in traffic to certain pages by
avoiding an unnecessary call to the Content Store. On the other hand, it
does mean that newly published content won't necessarily be visible
immediately, which is why I've set the cache time to a low 30 seconds.

There is a wider architectural question in this app related to when it
fetches from the content store. At the moment for each request there
will be two calls to the content store. First in the routes from the
`ContentFormatInspector` and `FormatRoutingConstraint` and then a second
time in the controller which is handling the request. It feels like this
could be simplified to only attempt to fetch from the content store
once however this is a bigger change as unpicking the bits of code will
take time. This new `CachedContentItem` solves the problem by caching,
but it doesn't actually stop the code from attempting to retrieve the
content item twice.

I based this code off similar code we've got in Finder Frontend:
https://github.com/alphagov/finder-frontend/blob/806aa066879c201fdbd9f9c3ab13331be2b4e143/app/lib/services.rb#L10-L16