Only make one request to the content store
This is quite a large commit, but all the bits of code here interact
very closely together and the app would be completely broken if it was
split up any further.

Ultimately this changes the controllers for most of the routes to use
the existing content item which was fetched as part of the routing
process to match the content item `schema_name` with the approriate
controller. This avoids making two requests to the content store for
each request.

I wasn't able to do this in a middleware because the `slug` of the route
needs to be known to fetch the item from the content store, and this
only happens once the routes have been parsed. Instead I've slightly
modified the `FormatRoutingConstraint` to store the content item in the
request environment, and then updated the `Navigable` concern to pull
that content item out of the request environment if it's there.

I've also taken the opportunity to rename the keys we're using to store
these in the request environment so they're slightly easier to reason
about: `:content_item` and `:content_item_error`.