Remove back links on local_transaction pages
What

There are both a back link and breadcrumbs on the results and unavailable_service
pages used by local_transactions. This is not a recommended pattern and one of
these elements should therefore be removed.

It is felt that we should remove the back link on these pages as breadcrumbs offer
a better user experience.

Remove the back links and update the tests.

Why

So that these pages are consistent with other GOV.UK pages and are in-line with
the design system guidance on the use of the back link component.