Refactor local transactions services model
This removes the need for a singleton class by creating an object per
service, which allows simplification of the code as the lgsl and country
name can become instance instance attributes. This also moves the config
to the lib folder, as the top level config is reserved for Rails
configuration.