Fix code coverage report when running rake default task
This was added in 94fdcd62ad54bd22fbc3539908231fa0c2314b77,
initially to track coverage across both Minitest and RSpec. The
latter has now been removed from Frontend, but I found that the
default 'test' task which Minitest gives inconsistent code
coverage results, based on the order in which the tasks are run.

'test' seems to run asynchronously - this is evidenced by, for
example, adding a `task :foo { puts 'bar' }` and changing the
default task to call `:foo` immediately after `:test` (we get the
output of `foo` before we get output of `test`).

I found that if I ran the `test` task in isolation, we had good
coverage:
> Coverage report generated for Minitest, Unit Tests to /govuk/frontend/coverage. 1243 / 1277 LOC (97.34%) covered.

If we ran the lint task before the test task, the coverage was
reduced:
> Coverage report generated for Minitest, Unit Tests to /govuk/frontend/coverage. 1233 / 1277 LOC (96.55%) covered.

And if we had the lint, test and Jasmine tests, the coverage was
reduced further (and 'Unit tests' removed from the summary below):
> Coverage report generated for Minitest to /govuk/frontend/coverage. 1118 / 1355 LOC (82.51%) covered.

I tried changing the ordering, to no avail. The only thing that
works is explicitly defining our own unit test task, which allows
us to run our tasks in any order and get consistent code coverage
results.