Modernise Local Transaction results view
This sets the font size in text to be a normal one rather than the
rather giant one that previously exited. It also includes more
supplementary information.

This changes the start button to be the less bold "start" button style
to be the more conventional success operation style. The text has been
changed to include the name of the local authority. This resolves in
accessibility issue that has been flagged in the Digital Accessibility
Centre's WCAG 2.1 audit of GOV.UK:

Screen reader comments:
“The button to move to the service of the relevant council announced to me as ‘Go to their
website’. It may be unclear to some screen reader users browsing out of context which
council the button relates to as the user may not have located the text situated within the
main content when browsing in context. I further would have expected to locate the feature
as a link, as a link would usually take the user to a different page or service. It would benefit
users if the item could be marked up as a link and could possess the name of the Council
within the link text to ensure screen reader users both in and out of context are aware of
what is being selected.”

Solution
Consider implementing an more succinct label, that accurately conveys the function or
purpose of the ‘Go to their website’ button to prevent any unnecessary confusion.
Example:

<a class="gem-c-button govuk-button govuk-button--start gem-c-button--bottom-
margin" role="button" rel="external" href="https://www.npt.gov.uk/" >
Go to Neath Port Talbot County Borough Council website
</a>