Add duplicate service with correct LGSL code
What

We want to add a duplicate configuration record for the lateral flow
mass testing finder service with a code of 1827. Once the service's LGSL
code has been switched over, the service with the LGSL code of 100001
will be removed.

Why

When the lateral flow mass testing finder service was created, it was
created with a dummy code of 100001.  We now have a real code (1827) so
would like to use this code instead of the dummy one.