Convoluted approach to change local transaction error copy
So I was tasked with changing the error copy for the local transaction
results page. This seemed like it should be really easy however it
wasn't as the handling of this is all wrapped up in a wider error
system. The designed change for this involved using a second paragraph
of an error (one explaining the area, the other the error) and this
wasn't compatible with the existing system.

To resolve this I've decided that there weren't any benefits of using
the wider error system for this scenario and I've simplified the code to
do the conditionals in the view.

There were a couple of unnecessary conditionals in the controller that
I've removed as they seemed to increase complexity and weren't tested
scenarios.

The first of these is the `if interaction_details["local_authority"]` -
I couldn't find any case where Local Links Manager could respond without
this information, and if it did this app would raise an error anyway
(due to accessing country name on a nil local_authority object).

The second was the removal of a guard clause for an empty local
authority slug. This didn't seem a possible scenario as you couldn't be
in the respective action without this slug.