Rename Top Actions section
This PR relates to special work we're doing to promote voting prior to the
local elections on 6 May.

We are not able to use a Top Actions tagged section, as we've been advised
that this now has to be visible in order to be eligible for use in a result.

We're renaming to "Related content" as this is more suited to the actual
use, and may be picked up for a different treatment in search result pages.