Add SessionsController, AccountConcern and Service
There are all being migrated across from finder-frontend. Before this
point we had run a discovery through the GOV.UK Accounts team to see if
we could experimentally add a login to just the brexit checker.

The discovery now wants to move on to a wider set of challenges, so it
makes sense to move the code out of finder-frontend to frontend, where
it can be a reference point for many services, not just the brexit
checker.

The code is mostly the same as it was in finder-frontend with a few
changes.

Service module:
==============

[see finder-frontend version][2]

In finder frontend this was a module that abstracted away the interfaces
for various other GOV.UK services. Here we can make a simple (and
shorter) call to just GdsApi.account_api. Which has replaced the
previous references to Services.account_api

[See here for previous example][3]

Account Concern
===============

[see finder-frontend version][4]

This is largely similar to the finder-frontend version.

However we were able to simplify the code by removing the need to handle
use cases where the AccountManager had been "turned off" via the feature
flag, or was not responding.

As we are now back in GOV.UK infrastructure we will want to re-examine
the plan there, and embed accounts availibility within our usual 2ndline
and on-call processes.

This means we can remove code that previously handled that in
finder-frontend

Session Controller
=================

[See finder-frontend version][5]

Again we have removed paths that handled disabled or unavailable
account apis, simplify the code but otherwise leaving it fairly
unchanged, with the exception of simplifying a test helper for the
redirect_with_ga method.

This was previously in place becuase we had some difficulty testing
feature tests in finder-frontend. That's not going to be a problem
here so we can simplify this a bit and let Minitest test for redirects
directly.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-134-govuk-wide-session-cookie-and-login.md
[2]: https://github.com/alphagov/finder-frontend/blob/release_2112/app/lib/services.rb#
[3]: https://github.com/alphagov/finder-frontend/blob/release_2112/app/controllers/sessions_controller.rb#L10
[4]: https://github.com/alphagov/finder-frontend/blob/release_2112/app/controllers/concerns/account_concern.rb
[5]: https://github.com/alphagov/finder-frontend/blob/release_2112/app/controllers/sessions_controller.rb

Co-authored-by: Alan Gabbianelli <alan.gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>