Add fetch_http_referer header
What
====

Requests to `/sign-in` should be sent here from buttons or links a user
has clicked somewhere on GOV.UK. With those requets should be an
HTTP_REFERER header that tells us where they were when they clicked.

Here we grab that header and use it to populate the redirect_path param
we send to the API.

Why
===

The Oauth2 flow allows a login and authentication journey to be embedded
within another feature our journey. So the user can continue down a path
(for instance filling out a form), then be redirected to authenticate
when there is a need.

These are interruptions to the journey, and so we will want to try and
return them to where they started. We prefer to do this via an explicit
redirect_path parameter that is attached to the URL of any links they
click.

The exception is the global header which will appear everywhere via
static, and so cannot know which URL it was previously on.

In this case, we will use a fallback method of trying to infer the users
starting location from an HTTP Referer header.

This is a best attempt where there URL parameter is not present, however
there may also be cases where it is missing, perhaps because of browser
privacy settings or an extension. In which case the user will be
directed to their account dashboard at the end of their journey.

How
===

Referer is User input and so cannot be fully trusted. We default to not
sending it if anything looks suspicious. In that case the user will be
taken to a fixed location like their account dashboard. We test for:

- If the referer header is present
- If the referer header starts with the Website root from plek
(https://gov.uk on production)
- If the result is a well formatted URI that can be parsed and have a
path returned.

If all is well there, we will populate the redirect_path param we send
to the API.

On line 9, we also accept a param[:redirect_path] and prefer this to
fetching HTTP headers. Meaning we can still add params to sign-in links
if we want to.

These are preferred to the header option as they are more explicit and
transparent to the user, so we will try for those first and only try for
headers after that.