Remove target="_blank" from postcode link
Removing this resolves an accessibility problem: that users are not
informed that a new tab will open.

The use of target=_blank is inconsistent for GOV.UK as a whole and
should only be applied mid-transaction [1]. This is not a mid
transaction scenario.

This also sets the URL to https since that is now the canonical URI for
the resource.

[1]: https://twitter.com/GDSTeam/status/983992989650423809