Adds country notice to the business support smart answer
We get the blue banner up (and down) on the business support smart answer for two weeks for the purpose of the test.

We are doing a before and after test and the banner is the key.

Styling and padding has been copied over from the metadata component however we have changed the font size to 19 to match start pages.

A time freeze has been added to the tests to ensure that the tests do not start failing after the experimental date

https://trello.com/c/Py8XjzRQ/275-implement-blue-banner-for-before-and-after-test