Add integration test for updating .home-numbers
We have this information stored in the organisations api, we can get the
data from there instead of manually updating each time.

Collections, which renders the organisations page seems to add a filter
that prevents "joining" organisations from being shown.[1]

As the point of this PR is to stop the numbers disagreeing, we should
ensure the logic matches here too.

[1]: https://github.com/alphagov/collections/blob/master/app/presenters/organisations/index_presenter.rb#L51-L55