Configure app to use memcached
Frontend currently uses the Rails default cache implementation,
which is to store entries to the file system at "#{root}/tmp/cache/".
We want it to use memcached instead.
Without setting the URL of the memcached servers, it will default to
localhost:11211, an additional puppet PR is needed to set the URL
as an env var.