Rename 'GDS_API_ADAPTERS_PACT_VERSION' => 'PACT_CONSUMER_VERSION'
This is more generic and allows us to standardise our Pact setup
regardless of what the Consumer and Provider is. Most of our Pact
tests have GDS Api Adapters as the Consumer, but we also have
tests between Content Store and Publishing API, where Publishing
API is the Consumer.
So rather than defining two different ENV variables and having
to account for the difference, the hope is that one common ENV
will allow us to consolidate as much behaviour as possible,
perhaps in somewhere like govuk_test.