Minor adjustments to the results page
An attempt to bring this closer to how we present addresses across gov.uk via govspeak. More on it https://components.publishing.service.gov.uk/component-guide/govspeak/address. It requires an override to prevent browsers from styling the address in italics, which would ideally be moved in govspeak. Aware there is some unstructured/unpredictable data, such as the email field or the date format in the upcoming elections.