Add lang attributes to bank holidays page
Currently, the Welsh version of the bank holidays page has lang='en'
even though most of the main content of the page is in Welsh.

These changes add a lang attribute to the main landmark on the page, so
that the content is correctly labelled as Welsh when viewing the Welsh
version of this page.
Additionally, when viewing the Welsh version, add a lang=en attribute to
the bank holidays experiment panel as we do not have translations for
most of this content. This is to ensure that assistive technologies can
read this content in the correct language, which is a WCAG requirement.