Change Visas & Immigration blurb
CHANGED
Visas, asylum and sponsorship
TO
Apply to visit, work, study, settle or seek asylum in the UK

REASON
This will help people who don't understand the words 'visas' and 'immigration' to understand better what is in this browse section. 
The new description also matches the new structure of the Visas and immigration browse section.