Always send the user to authenticate from /sign-in
Now that we have multiple levels of authentication, the user can be
logged in and still get sent to the /sign-in path.  Rather than try to
do anything clever (like extracting the current level of
authentication from the session and comparing it with the desired),
just always send the user to auth.  If they have a valid session for
the account manager they'll get redirected back to the callback path
immediately.

This also fixes an inconsistency where

    GET /sign-in?redirect_path=/foo

Would send you to the account manager if logged in; but if not
authenticate you and send you to www.gov.uk/foo.