Fix JS crossbrowser compatibility issue
Swap the replaceAll() string method for the replace() method with a
global flag on the bank holidays save nation script.
The replaceAll() method is not supported in any version of IE which is
causing an error. As a consequence, on Internet Explorer the save nation
panel currently reads "Do you want to save # as your location for bank
holidays" instead of e.g "Do you want to save Scotland as your location
for bank holidays".