Move account methods up to ApplicationController
Previously these were only needed for the getting and setting GOV.UK
Account sessions. However frontend is now also the place we'll be
relaying account feature requests through (specifically save a page).

To prevent rewriting similar concerns I've moved these up to the
ApplicationController for the moment.

However they are unlikely to stay there for long!

The work on the [govuk_personalisation gem][1] should shortly supercede
this commit. At which point these methods should be available from
within the shared concern via the gem.

If that gets merged before this PR, I'll rebase and drop this commit.
If this makes it there first we'll go back and tweak this when we are
integrating the Gem.

[1]: https://github.com/alphagov/govuk_personalisation