Add /account/saved-pages/add
/account/saved-pages/new is an authenticated path, if the user attempts
to save a page without an account_session_header then logged_in? will be
false and we should redirect them to authenticate.

The save page feature only requires level0 auth (No MFA needed!)
We can pass on the state_id, and should redirect them back to the save a
page path so they can try again after auth.

Add /account/saved-pages/new - logged in behaviour

If a user is authenticated then they can save a page!

I have added a rescue in here, I will check in the review if it is
really required. I was considering a scenario where the
account_session_header is populated (perhaps with hacker junk) and the
API returns an unauthorised.

In that case seems best to clear the session header and attempt to
re-authenticate from scratch

ToSquash: Protect against redirect attacks

Brakeman has rightly noted that we were offering a redirect based on an
unverified parameter. This was a good catch.

For the Happy path there is not really an issue here, the account-api
will have gone and confirmed a path represents an actually existing bit
of content on GOV.UK, meaning that the redirect path we send back is a
valid path within GOV.UK.

However, if we're getting back a 422 error then the user could have sent
anything across, and in this case we have an obligation to ensure that
there are no redirect attacks that would allow a user to be sent a
seemingly valid GOV.UK domain link (with a sneaky param) that then
redirects them back to a separate domain and evil site.

ToSquash: Gate /account/saved-pages/new behind a feature flag