Return a 422 error when saving a page if the account-api does
account-api returning a 422 means that the user tried to save an
invalid path.

If this happens to a legitimate user (i.e., they haven't tampered with
the URL trying to break our stuff) then this is a bug in whatever page
had the button, in which case we *want* users to open support tickets
saying "I tried to save a page but got an error!"

If this happens to a user who has tampered with the URL, it's their
own fault.

We might want to return a `page_not_saved` flag to the page if
account-api returns a 5xx error code, because that likely indicates a
transient problem.  Though in that case we should do some more path
validation in this app.