Prefer font weight with css over use of <strong>
We prefer to do styling in CSS rather than rely on HTML defaults.
In addition the spec as [described by MDN][1] suggests that if you _are_
going to use it in HTML it should have some kind of semantic meaning:

> The <strong> element is for content that is of "strong importance,"
> including things of great seriousness or urgency (such as warnings).
> This could be a sentence that is of great importance to the whole page,
> or you could merely try to point out that some words are of greater
> importance compared to nearby content.

Given these figures likely have a faded importance since the majority of
government websites are now sucesfully migrated to GOV.UK, I think it's
probably no longer the a central figure with great seriousness or
urgency.

So... let's switch it to a <span>

In addition app/assets/stylesheets/views/_homepage.scss:L186 already
sets a `font-weight: bold;` within the class `.home-numbers__large`. So
the styling was already being done in CSS anyway!

[1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/strong#usage_notes