Resolve missing .govuk-grid-row for calendar views
These views were using the .govuk-grid-column-* classes without them
being inside a .govuk-grid-row which meant that the column floats
weren't cleared.

This caused the following page component, the feedback prompt, to share
the same layout as the columns. This problem is barely noticeable when
Javascript is enabled - with just some minor spacing problems at the
footer. However without javascript it caused styling to spill out making
the page look broken.