Determine devolved administration availability for a transaction
This checks the devolved administration availability for a service based on the country name that is returned from the postcode lookup.

This will be used by the controller to determine which results page to display to the user (in a following commit).

It also sets the devolved_administration_service_alternative_url to return the alternative_url provided for a devolved_administration_service, which will be accessed from the relevant view in a future commit.

Previously the availability and button_link text were set in the LocalTransactionService. This functionality has been moved to the presenter and the LocalTransactionService will be deleted in a following commit. As these methods determine how information is presented to a user on the view, a presenter seemed like a better place for this functionality.

I have chosen a more traditional approach of setting the text in the views and rendering them separately, rather
than using one view and displaying different text based on what the methods return (as was the case previously in LocalTransactionService). I felt this approach was easier to follow and didn’t really involve too much duplication.