Views for displaying results to user
Views are rendered for a user depending on devolved_administration_availability.

The more_information section is not rendered from the devolved_administration_service or the unavailable_service view. This is due to concerns over the content not being relevant. However it can be added in at a later date if requested by content designers.

The analytics values have been agreed with a performance analyst (Harriet Wailling).

Rationale for the chosen values (taken from a conversation between Harriet and Kevin Dew) -

For a service that is handled differently by a devolved administration:

It is a form of result (we matched their postcode to a country) but is different to a normal result so we’ll give it a different name. Use country to identify which countries are accessed most (for existing local administration results we track a label of the local authority name).

For a service that is unavailable to a devolved administration:

Again a form of result since a match of postcode to country. Having the postcodeSearch:local_transaction event fire regardless of whether a local authority link exists or not is a convention already in place for existing local administration results.

When we don’t have a local authority link for an unavailable result:

This records a userAlert for a local transaction rather than a postcode search and this alert is the same whether service is unavailable or not (and we already have the other event that distinguishes). Harriet and I also felt that this event was so rare and barely used so it wasn’t worth doing a lot of new thinking for it.